requirements.txt
Django~=4.1.2
djangorestframework~=3.14.0
celery~=5.2.7
python-dotenv~=0.21.1
pytest~=7.2.2
pytest-django~=4.5.2
redis~=4.5.2
drf-yasg~=1.21.5
djoser~=2.1.0
psycopg2~=2.9.5
djangorestframework-simplejwt~=4.8.0
social-auth-app-django~=4.0.0
flake8~=6.0.0
django_redis~=5.2.0